+++ packages/frontend/src/app/a/[slug]/page.tsx
@@ -4,6 +4,8 @@ import { notFound } from 'next/navigation'
 import SpeechPage from '@/components/speech'
 // fetcher
 import { fetchSpeech, fetchSpeechGroup } from '@/fetchers/server/speech'
+// constants
+import { InternalRoutes } from '@/constants/routes'
 
 export const dynamicParams = true
 
@@ -20,9 +22,19 @@ export async function generateMetadata({
   const { title, legislativeYuanMember } = speech
   const { legislator } = legislativeYuanMember
   const { name } = legislator
+  const titleForMetaData =
+    title.length > 15 ? `${title.slice(0, 15)}...` : title
   return {
-    title: `逐字稿 - ${name} - ${title}`,
-    description: `議會逐字稿 - ${title}`,
+    title: `逐字稿｜${name}：${titleForMetaData} - 報導者觀測站`,
+    description: '報導者議會透視版',
+    alternates: {
+      canonical: `https://lawmaker.twreporter.org${InternalRoutes.Speech}/${slug}`,
+    },
+    openGraph: {
+      title: `逐字稿｜${name}：${titleForMetaData} - 報導者觀測站`,
+      url: `https://lawmaker.twreporter.org${InternalRoutes.Speech}/${slug}`,
+      type: 'article',
+    },
   }
 }