+++ packages/frontend/src/components/open/config.ts
@@ -1,4 +1,4 @@
-export const title = '議會透視板'
+export const title = '報導者觀測站'
 export const description = [
   '你家立委關心什麼？你關注的議題有誰在討論？',
   '點擊你感興趣的人或話題，觀察立委的發言與提案情形',